Moved License code to separate app
