+++ FOSSDB_web/apps/fossdb/models.py
@@ -3,15 +3,7 @@ import uuid
 from django.contrib.auth.models import User
 from django.db import models
 
-
-class License(models.Model):
-    short_name = models.CharField(max_length=50)
-    full_name = models.CharField(max_length=100, null=True, blank=True)
-    url = models.URLField(null=True, blank=True)
-    description = models.TextField(null=True, blank=True)
-
-    def __str__(self):
-        return self.short_name
+from license.models import License
 
 
 class ProgrammingLanguage(models.Model):
