+++ FOSSDB_web/apps/license/models.py
@@ -1,3 +1,11 @@
 from django.db import models
 
-# Create your models here.
+
+class License(models.Model):
+    short_name = models.CharField(max_length=50)
+    full_name = models.CharField(max_length=100, null=True, blank=True)
+    url = models.URLField(null=True, blank=True)
+    description = models.TextField(null=True, blank=True)
+
+    def __str__(self):
+        return self.short_name
